Wrap lottoNumberPool list in LottoNumberPool class
